split policies into individual files
At >100 lines, the file was becoming a bit unwieldy.
