+++ src/constructs/iam/policies/index.ts
@@ -1,0 +1,4 @@
+export * from "./base-policy";
+export * from "./log-shipping";
+export * from "./s3-get-object";
+export * from "./ssm";
