+++ src/constructs/iam/policies/log-shipping.ts
@@ -1,0 +1,32 @@
+import type { PolicyProps } from "@aws-cdk/aws-iam";
+import { Effect, PolicyStatement } from "@aws-cdk/aws-iam";
+import type { GuStack } from "../../core";
+import type { GuPolicyProps } from "./base-policy";
+import { GuPolicy } from "./base-policy";
+
+export interface GuLogShippingPolicyProps extends GuPolicyProps {
+  kinesisStreamName: string;
+}
+
+// This interface is inconsistent. Should we add the ID here?
+export class GuLogShippingPolicy extends GuPolicy {
+  private static getDefaultProps(scope: GuStack, props: GuLogShippingPolicyProps): PolicyProps {
+    return {
+      policyName: "log-shipping-policy",
+      statements: [
+        new PolicyStatement({
+          effect: Effect.ALLOW,
+          actions: ["kinesis:Describe*", "kinesis:Put*"],
+          resources: [`arn:aws:kinesis:${scope.region}:${scope.account}:stream/${props.kinesisStreamName}`],
+        }),
+      ],
+    };
+  }
+
+  constructor(scope: GuStack, id: string = "LogShippingPolicy", props: GuLogShippingPolicyProps) {
+    super(scope, id, {
+      ...GuLogShippingPolicy.getDefaultProps(scope, props),
+      ...props,
+    });
+  }
+}
